separating out csharp generator logic
